@@ -22,6 +22,10 @@
     {
       "name": "Benjamin Saphier",
       "url": "https://github.com/bsaphier"
+    },
+    {
+      "name": "James Robb",
+      "url": "https://github.com/jamesrweb"
     }
   ],
   "homepage": "https://github.com/and-who/react-p5-wrapper",
@@ -34,27 +38,27 @@
     "url": "https://github.com/and-who/react-p5-wrapper/issues"
   },
   "devDependencies": {
-    "@babel/cli": "^7.4.4",
-    "@babel/core": "^7.4.5",
-    "@babel/preset-env": "^7.4.5",
-    "@babel/preset-react": "^7.0.0",
-    "@types/p5": "^0.7.2",
-    "@types/react": "^16.8.18",
-    "@types/react-dom": "^16.8.4",
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.9.6",
+    "@babel/preset-env": "^7.9.6",
+    "@babel/preset-react": "^7.9.4",
+    "@types/p5": "^0.7.3",
+    "@types/react": "^16.9.35",
+    "@types/react-dom": "^16.9.8",
     "awesome-typescript-loader": "^5.2.1",
-    "babel-loader": "^8.0.6",
+    "babel-loader": "^8.1.0",
     "css-loader": "^2.1.1",
-    "gh-pages": "^2.0.1",
+    "gh-pages": "^2.2.0",
     "html-webpack-plugin": "^3.2.0",
     "p5": "^0.8.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "source-map-loader": "^0.2.4",
     "style-loader": "^0.23.1",
-    "typescript": "^3.4.5",
-    "webpack": "^4.32.2",
-    "webpack-cli": "^3.3.2",
-    "webpack-dev-server": "^3.4.1"
+    "typescript": "^3.9.3",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.11.0"
   },
   "peerDependencies": {
     "react": "^16.3.0",
