@@ -1,37 +1,32 @@
-import React from'react';
-import ReactDOM from 'react-dom';
-import P5Wrapper from '../../src';
-import sketch from './sketches/sketch';
-import sketch2 from './sketches/sketch2';
-import css from './example.css';
+import React, { Fragment, useState } from "react";
+import { render } from "react-dom";
+import P5Wrapper from "../../src";
+import sketch from "./sketches/sketch";
+import sketch2 from "./sketches/sketch2";
+import css from "./example.css";
 
-class App extends React.Component{
+function App(props) {
+	const [state, setState] = useState({ rotation: 160, sketch });
 
-	constructor(props) {
-		super(props);
-		this.state = {
-			rotation: 150,
-			stateSketch: sketch,
-		};
-	}
-
-	rotationChange(e){
-		this.setState({rotation:e.target.value});
-	}
-
-	pressEvent(){
-		this.state.stateSketch === sketch ? this.setState({stateSketch:sketch2}) : this.setState({stateSketch:sketch});
-	}
-
-	render () {
-		return (
-			<div>
-				<P5Wrapper sketch={this.state.stateSketch} rotation={this.state.rotation}/>
-				<input type="range" value={this.state.rotation}  min="0"  max="360" step="1" onInput={this.rotationChange.bind(this)}/>
-				<button onClick={this.pressEvent.bind(this)}>Change Sketch</button>
-			</div>
-		);
-	}
+	return (
+		<Fragment>
+			<P5Wrapper sketch={state.sketch} rotation={state.rotation} />
+			<input
+				type="range"
+				defaultValue={state.rotation}
+				min="0"
+				max="360"
+				step="1"
+				onChange={event => setState({ ...state, rotation: event.target.value })}
+			/>
+			<button onClick={
+				event => setState({
+					...state,
+					sketch: state.sketch === sketch ? sketch2 : sketch
+				})
+			}>Change Sketch</button>
+		</Fragment>
+	);
 }
 
-ReactDOM.render(<App />, document.getElementById('app'));
+render(<App />, document.getElementById("app"));
