@@ -1,65 +1,74 @@
-import { Component } from 'react';
-import p5 = require('p5');
-import React = require('react');
-import { stat } from 'fs';
+import p5 from 'p5';
+import React from 'react';
 
-export interface P5WrapperProps {
+export interface IP5WrapperProps {
   sketch: (p: p5) => void;
 }
-export interface P5WrapperState {
+
+export interface IP5WrapperState {
   sketch: (p: p5) => void;
   canvas: p5;
-  wrapper: HTMLElement;
+  wrapper?: HTMLElement;
 }
 
-export default class P5Wrapper extends Component <P5WrapperProps, P5WrapperState> {
+class P5Wrapper extends React.Component<IP5WrapperProps, IP5WrapperState> {
+  public state: IP5WrapperState;
+  public wrapper: HTMLElement = null;
 
-  constructor(props: any) {
+  constructor(props: IP5WrapperProps) {
     super(props);
-    this.state  = {
-      sketch: props.sketch,
+    this.state = {
+      ...props,
       canvas: null,
-      wrapper: null
-    };
+      // TODO: find a workaround since refs in state are bad practice
+      wrapper: this.wrapper
+    }
   }
 
-  wrapper: HTMLElement;
+  static getDerivedStateFromProps(
+    props: IP5WrapperProps,
+    state: IP5WrapperState
+  ) {
+    if (state.sketch !== props.sketch) {
+      const { sketch } = props;
+      const canvas = new p5(sketch, state.wrapper);
+      state.canvas.remove();
+      return { ...state, sketch, canvas };
+    }
+
+    // @ts-ignore
+    // @TODO: fix type definitions
+    if (state.canvas && state.canvas.myCustomRedrawAccordingToNewPropsHandler) {
+      // @ts-ignore
+      // @TODO: fix type definitions
+      state.canvas.myCustomRedrawAccordingToNewPropsHandler(props);
+    }
+
+    return state;
+  }
 
   componentDidMount() {
     const canvas = new p5(this.state.sketch, this.wrapper);
-    if( canvas.myCustomRedrawAccordingToNewPropsHandler ) {
+    // @ts-ignore
+    // @TODO: fix type definitions
+    if (canvas.myCustomRedrawAccordingToNewPropsHandler) {
+      // @ts-ignore
+      // @TODO: fix type definitions
       canvas.myCustomRedrawAccordingToNewPropsHandler(this.props);
     }
-    this.setState({
-      canvas: canvas,
-      wrapper: this.wrapper
-    })
-  }
-
-  static getDerivedStateFromProps(props: P5WrapperProps, state: P5WrapperState){
-    let canvas = state.canvas;
-    if(state.sketch !== props.sketch){
-      state.wrapper.removeChild(state.wrapper.childNodes[0]);
-      canvas.remove();
-      canvas = new p5(props.sketch, state.wrapper)
-      return {
-        ...state,
-        sketch: props.sketch,
-        canvas: canvas
-      };
-    }
-    if( canvas && canvas.myCustomRedrawAccordingToNewPropsHandler ) {
-      canvas.myCustomRedrawAccordingToNewPropsHandler(props);
-    }
-    return state;
+    this.setState({ ...this.state, canvas, wrapper: this.wrapper });
   }
 
 
   componentWillUnmount() {
-		this.state.canvas.remove();
-	}
+    if (this.state.canvas !== null) {
+      this.state.canvas.remove();
+    }
+  }
 
   render() {
     return <div ref={wrapper => this.wrapper = wrapper}></div>;
   }
 }
+
+export default P5Wrapper;
